Restarted kernel space implementation

+++ include/linux/syscalls.h
@@ -1383,7 +1383,7 @@ int __sys_setsockopt(int fd, int level, int optname, char __user *optval,
 		int optlen);
 
 asmlinkage long sys_add_task(queue_t* q, task_t t);
-asmlinkage long sys_schedule_tasks(void);
+asmlinkage int sys_schedule_tasks(void);
 asmlinkage long sys_acquire_resources(resource_t r[2], task_t t);
 asmlinkage long sys_release_resources(resource_t r[2], task_t t);
 #endif
